Update Makefile to work with go mod
